+++ bin/generate_dax.sh
