+++ bin/hcp.sh
