+++ bin/plan_dax.sh
