+++ bin/seeg-ct.sh
