+++ bnm/recon/daxgen/__main__.py
@@ -1,0 +1,4 @@
+
+from . import freesurfer_main
+
+freesurfer_main()
