+++ docs/setup-vm/00-system.sh
