+++ docs/setup-vm/10-cmake.sh
