+++ docs/setup-vm/20-gsl.sh
