+++ docs/setup-vm/25-mrtrix.sh
