+++ docs/setup-vm/30-hdf5.sh
