+++ docs/setup-vm/35-openmeeg.sh
