+++ docs/setup-vm/50-freesurfer.sh
