+++ docs/setup-vm/60-mne.sh
