+++ docs/setup-vm/setup-software.sh
