+++ docs/setup-vm/setup-system-and-reboot.sh
