+++ docs/setup-vm/setup-user-env.sh
