--- engine/bnm.dax
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- generated: 2016-10-04 10:11:41.614293 -->
-<!-- generated by: lia.domide -->
-<!-- generator: python -->
-<adag xmlns="http://pegasus.isi.edu/schema/DAX" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://pegasus.isi.edu/schema/DAX http://pegasus.isi.edu/schema/dax-3.6.xsd" version="3.6" name="BrainNetworkModelReconstructionFlow">
-	<metadata key="created">Tue Oct  4 10:11:41 2016</metadata>
-	<job id="ID0000001" name="mri_convert" node-label="T1 input pre-processing">
-		<argument><file name="/WORK/VEP/JUNG/T1"/> <file name="/WORK/VEP/JUNG/T1/t1_0.038104.nii.gz"/> --out_orientation RAS -rt nearest</argument>
-		<uses name="/WORK/VEP/JUNG/T1" link="input"/>
-		<uses name="/WORK/VEP/JUNG/T1/t1_0.038104.nii.gz" link="output" register="false" transfer="false"/>
-	</job>
-	<job id="ID0000002" name="recon-all" node-label="Call &apos;recon-all&apos;">
-		<argument>-s JUNG -i <file name="/WORK/VEP/JUNG/T1/t1_0.038104.nii.gz"/> -all -parallel -openmp 2</argument>
-		<uses name="/WORK/VEP/JUNG/T1/t1_0.038104.nii.gz" link="input"/>
-	</job>
-	<job id="ID0000003" name="mri_convert" node-label="Generate nifti file with good orientation">
-		<argument><file name="/WORK/VEP/JUNG/MRI/aparc+aseg.mgz"/> <file name="/WORK/VEP/JUNG/MRI/aparc+aseg_0.038218.nii.gz"/> --out_orientation RAS -rt nearest</argument>
-		<uses name="/WORK/VEP/JUNG/MRI/aparc+aseg.mgz" link="input"/>
-		<uses name="/WORK/VEP/JUNG/MRI/aparc+aseg_0.038218.nii.gz" link="output" register="false" transfer="false"/>
-	</job>
-	<child ref="ID0000002">
-		<parent ref="ID0000001"/>
-	</child>
-	<child ref="ID0000003">
-		<parent ref="ID0000002"/>
-	</child>
-</adag>
