+++ extern/TODO.md
@@ -1,0 +1,5 @@
+Preferably Git submodules, tracking specific versions
+
+- pegasus (or at least Python lib)
+- openmeeg
+- mrtrix3
